+++ app/controllers/metadata_presenter/answers_controller.rb
@@ -4,7 +4,7 @@ module MetadataPresenter
 
     def create
       @previous_answers = reload_user_data.deep_dup
-      @page_answers = PageAnswers.new(page, answers_params)
+      @page_answers = PageAnswers.new(page, answers_params, autocomplete_items(page.components))
 
       upload_files if upload?
 
@@ -43,6 +43,8 @@ module MetadataPresenter
 
     def render_validation_error
       @user_data = answers_params
+      load_autocomplete_items
+
       render template: page.template, status: :unprocessable_entity
     end