+++ app/validators/metadata_presenter/validate_answers.rb
@@ -1,10 +1,11 @@
 module MetadataPresenter
   class ValidateAnswers
-    attr_reader :page_answers, :components
+    attr_reader :page_answers, :components, :autocomplete_items
 
-    def initialize(page_answers, components:)
+    def initialize(page_answers, components:, autocomplete_items:)
       @page_answers = page_answers
       @components = Array(components)
+      @autocomplete_items = autocomplete_items
     end
 
     def valid?
@@ -21,8 +22,10 @@ module MetadataPresenter
       components.map { |component|
         component_validations(component).map do |key|
           "MetadataPresenter::#{key.classify}Validator".constantize.new(
-            page_answers: page_answers,
-            component: component
+            {
+              component: component,
+              page_answers: page_answers
+            }.merge(autocomplete_param(key))
           )
         end
       }.compact.flatten
@@ -32,6 +35,10 @@ module MetadataPresenter
       return [] if component.validation.blank?
 
       component.validation.select { |_, value| value.present? }.keys
+    end
+
+    def autocomplete_param(key)
+      key == 'autocomplete_item' ? { autocomplete_items: autocomplete_items } : {}
     end
   end
 end