+++ spec/validators/validate_answers_spec.rb
@@ -1,8 +1,9 @@
 RSpec.describe MetadataPresenter::ValidateAnswers do
   subject(:validate_answers) do
-    described_class.new(page_answers, components: page.components)
+    described_class.new(page_answers, components: page.components, autocomplete_items: autocomplete_items)
   end
   let(:page_answers) { MetadataPresenter::PageAnswers.new(page, answers) }
+  let(:autocomplete_items) { {} }
   let(:page) { service.find_page_by_url('/name') }
 
   describe '#valid?' do